resetting project with boilerplate ts
