--- .eslintcache
@@ -1,1 +1,0 @@
-[{"/Users/ericmahoney/Documents/programming/react-course/musicapp/src/index.js":"1","/Users/ericmahoney/Documents/programming/react-course/musicapp/src/App.js":"2","/Users/ericmahoney/Documents/programming/react-course/musicapp/src/reportWebVitals.js":"3","/Users/ericmahoney/Documents/programming/react-course/musicapp/src/components/Library.js":"4","/Users/ericmahoney/Documents/programming/react-course/musicapp/src/components/Player.js":"5","/Users/ericmahoney/Documents/programming/react-course/musicapp/src/components/CurrentSong.js":"6","/Users/ericmahoney/Documents/programming/react-course/musicapp/src/components/Utilities.js":"7","/Users/ericmahoney/Documents/programming/react-course/musicapp/src/components/LibrarySong.js":"8","/Users/ericmahoney/Documents/programming/react-course/musicapp/src/components/Navbar.js":"9"},{"size":500,"mtime":1609385574648,"results":"10","hashOfConfig":"11"},{"size":1578,"mtime":1610318142041,"results":"12","hashOfConfig":"11"},{"size":362,"mtime":1609385574650,"results":"13","hashOfConfig":"11"},{"size":825,"mtime":1610314819770,"results":"14","hashOfConfig":"11"},{"size":3443,"mtime":1610314754418,"results":"15","hashOfConfig":"11"},{"size":544,"mtime":1610312304815,"results":"16","hashOfConfig":"11"},{"size":2624,"mtime":1610305611800,"results":"17","hashOfConfig":"11"},{"size":993,"mtime":1610314733298,"results":"18","hashOfConfig":"11"},{"size":379,"mtime":1610304352083,"results":"19","hashOfConfig":"11"},{"filePath":"20","messages":"21","errorCount":0,"warningCount":0,"fixableErrorCount":0,"fixableWarningCount":0,"usedDeprecatedRules":"22"},"vyyfy1",{"filePath":"23","messages":"24","errorCount":0,"warningCount":0,"fixableErrorCount":0,"fixableWarningCount":0},{"filePath":"25","messages":"26","errorCount":0,"warningCount":0,"fixableErrorCount":0,"fixableWarningCount":0,"usedDeprecatedRules":"22"},{"filePath":"27","messages":"28","errorCount":0,"warningCount":0,"fixableErrorCount":0,"fixableWarningCount":0},{"filePath":"29","messages":"30","errorCount":0,"warningCount":0,"fixableErrorCount":0,"fixableWarningCount":0},{"filePath":"31","messages":"32","errorCount":0,"warningCount":0,"fixableErrorCount":0,"fixableWarningCount":0},{"filePath":"33","messages":"34","errorCount":0,"warningCount":0,"fixableErrorCount":0,"fixableWarningCount":0},{"filePath":"35","messages":"36","errorCount":0,"warningCount":0,"fixableErrorCount":0,"fixableWarningCount":0},{"filePath":"37","messages":"38","errorCount":0,"warningCount":0,"fixableErrorCount":0,"fixableWarningCount":0},"/Users/ericmahoney/Documents/programming/react-course/musicapp/src/index.js",[],["39","40"],"/Users/ericmahoney/Documents/programming/react-course/musicapp/src/App.js",[],"/Users/ericmahoney/Documents/programming/react-course/musicapp/src/reportWebVitals.js",[],"/Users/ericmahoney/Documents/programming/react-course/musicapp/src/components/Library.js",[],"/Users/ericmahoney/Documents/programming/react-course/musicapp/src/components/Player.js",[],"/Users/ericmahoney/Documents/programming/react-course/musicapp/src/components/CurrentSong.js",[],"/Users/ericmahoney/Documents/programming/react-course/musicapp/src/components/Utilities.js",[],"/Users/ericmahoney/Documents/programming/react-course/musicapp/src/components/LibrarySong.js",[],"/Users/ericmahoney/Documents/programming/react-course/musicapp/src/components/Navbar.js",[],{"ruleId":"41","replacedBy":"42"},{"ruleId":"43","replacedBy":"44"},"no-native-reassign",["45"],"no-negated-in-lhs",["46"],"no-global-assign","no-unsafe-negation"]
