+++ src/App.css
@@ -1,58 +1,38 @@
-@import url("https://fonts.googleapis.com/css2?family=Nunito+Sans:wght@400;700&display=swap");
-
-* {
-  margin: 0;
-  padding: 0;
-  box-sizing: border-box;
+.App {
+  text-align: center;
 }
 
-body {
-  font-family: "Nunito Sans", sans-serif;
+.App-logo {
+  height: 40vmin;
+  pointer-events: none;
 }
 
-button {
-  border: none;
-  margin: 0;
-  padding: 0;
-  width: auto;
-  overflow: visible;
-  color: inherit;
-  font: inherit;
-  line-height: normal;
-  -webkit-font-smoothing: inherit;
-  -moz-osx-font-smoothing: inherit;
-  -webkit-appearance: none;
-  background-color: #023047;
-  color: white;
-  padding: 5px 10px;
-  border-radius: 5px;
+@media (prefers-reduced-motion: no-preference) {
+  .App-logo {
+    animation: App-logo-spin infinite 20s linear;
+  }
 }
 
-.app-container {
-  height: 100vh;
-  display: grid;
-  grid-template-areas:
-    "navbar"
-    "song"
-    "player";
-  grid-template-rows: 10% 1fr 20%;
-}
-
-.footer {
-  padding: 20px;
+.App-header {
+  background-color: #282c34;
+  min-height: 100vh;
   display: flex;
-  justify-content: flex-end;
-  width: 100%;
+  flex-direction: column;
+  align-items: center;
+  justify-content: center;
+  font-size: calc(10px + 2vmin);
+  color: white;
 }
 
-.footer a {
-  text-decoration: none;
-  color: black;
+.App-link {
+  color: #61dafb;
 }
 
-@media (max-width: 600px) {
-  .app-container {
-    height: 70vh;
-    grid-template-rows: 10% 1fr 20%;
+@keyframes App-logo-spin {
+  from {
+    transform: rotate(0deg);
+  }
+  to {
+    transform: rotate(360deg);
   }
 }
