--- src/components/Library.css
@@ -1,36 +1,0 @@
-.library-container {
-  opacity: 0;
-  position: fixed;
-  top: 0;
-  left: 0;
-  bottom: 0;
-  width: 20rem;
-  box-shadow: 0px 6px 50px rgba(0, 0, 0, 0.13);
-  background-color: white;
-  z-index: 999;
-  color: black;
-  overflow: scroll;
-  transform: translateX(-100%);
-  transition: all 0.5s ease;
-}
-
-.close-btn {
-  display: none;
-  position: absolute;
-  top: 20px;
-  right: 20px;
-}
-
-.active-library {
-  transform: translateX(0%);
-  opacity: 1;
-}
-
-@media (max-width: 800px) {
-  .library-container {
-    width: 100%;
-  }
-  .close-btn {
-    display: block;
-  }
-}
