--- src/components/Navbar.css
@@ -1,13 +1,0 @@
-.navbar-wrapper {
-  align-self: center;
-}
-
-.navbar {
-  display: flex;
-  justify-content: space-between;
-  padding: 0px 20px;
-}
-
-.logo {
-  font-weight: 400;
-}
