--- src/components/Navbar.js
@@ -1,16 +1,0 @@
-import "./Navbar.css";
-
-const Navbar = ({ setLibraryOpen, libraryOpen }) => {
-  return (
-    <div className="navbar-wrapper">
-      <nav className="navbar">
-        <h1 className="logo">Music App</h1>
-        <button className="library-btn" onClick={() => setLibraryOpen(!libraryOpen)}>
-          Library
-        </button>
-      </nav>
-    </div>
-  );
-};
-
-export default Navbar;
