--- src/components/Player.css
@@ -1,119 +1,0 @@
-.player-container {
-  grid-area: player;
-  width: 100%;
-  height: 100%;
-  align-self: center;
-}
-
-.player {
-  height: 100%;
-  display: flex;
-  flex-direction: column;
-  justify-content: space-evenly;
-}
-
-.player-controls {
-  margin: 0 auto;
-  display: flex;
-  justify-content: space-between;
-  width: 250px;
-  font-size: 2.5rem;
-}
-
-.timeline-container {
-  margin: 0 auto;
-  width: 500px;
-  display: flex;
-  justify-content: center;
-}
-
-.song-slider {
-  -webkit-appearance: none;
-  margin: 10px 0;
-  width: 100%;
-}
-
-.song-slider::-webkit-slider-runnable-track {
-  width: 100%;
-  height: 5px;
-  cursor: pointer;
-  box-shadow: 0px 0px 0px #000000;
-  background: #2497e3;
-  border-radius: 1px;
-  border: 0px solid #000000;
-}
-.song-slider::-webkit-slider-thumb {
-  box-shadow: 0px 0px 0px #000000;
-  border: 1px solid #2497e3;
-  height: 18px;
-  width: 18px;
-  border-radius: 25px;
-  background: #a1d0ff;
-  cursor: pointer;
-  -webkit-appearance: none;
-  margin-top: -7px;
-}
-.song-slider:focus::-webkit-slider-runnable-track {
-  background: #2497e3;
-}
-.song-slider::-moz-range-track {
-  width: 100%;
-  height: 5px;
-  cursor: pointer;
-  box-shadow: 0px 0px 0px #000000;
-  background: #2497e3;
-  border-radius: 1px;
-  border: 0px solid #000000;
-}
-.song-slider::-moz-range-thumb {
-  box-shadow: 0px 0px 0px #000000;
-  border: 1px solid #2497e3;
-  height: 18px;
-  width: 18px;
-  border-radius: 25px;
-  background: #a1d0ff;
-  cursor: pointer;
-}
-.song-slider::-ms-track {
-  width: 100%;
-  height: 5px;
-  cursor: pointer;
-  background: transparent;
-  border-color: transparent;
-  color: transparent;
-}
-.song-slider::-ms-fill-lower {
-  background: #2497e3;
-  border: 0px solid #000000;
-  border-radius: 2px;
-  box-shadow: 0px 0px 0px #000000;
-}
-.song-slider::-ms-fill-upper {
-  background: #2497e3;
-  border: 0px solid #000000;
-  border-radius: 2px;
-  box-shadow: 0px 0px 0px #000000;
-}
-.song-slider::-ms-thumb {
-  margin-top: 1px;
-  box-shadow: 0px 0px 0px #000000;
-  border: 1px solid #2497e3;
-  height: 18px;
-  width: 18px;
-  border-radius: 25px;
-  background: #a1d0ff;
-  cursor: pointer;
-}
-.song-slider:focus::-ms-fill-lower {
-  background: #2497e3;
-}
-.song-slider:focus::-ms-fill-upper {
-  background: #2497e3;
-}
-
-@media (max-width: 500px) {
-  .player-controls,
-  .timeline-container {
-    width: 80%;
-  }
-}
